requirements.txt
django>=3.0.3
djangorestframework>=3.11.0
django-filter>=2.2.0
django-rest-knox>=4.1.0
django-phonenumber-field>=4.0.0
psycopg2>=2.8.4
python-dotenv>=0.12.0
Pillow>=7.0.0
argon2-cffi>=19.2.0
django-cors-headers>=3.2.1
django-environ>=0.4.5
mozilla-django-oidc>=1.2.3